Store Tests: integrated creator methods and reduced duplication
