+++ test/store/custom_list/fixture.py
@@ -3,25 +3,25 @@
 from __future__ import annotations
 
 import typing
+from pathlib import Path
 from typing import List, Optional
 
 import pytest
+from sqlalchemy import StaticPool
 
 from mlte.backend.core.config import settings
 from mlte.custom_list.custom_list_names import CustomListName
 from mlte.custom_list.model import CustomListEntryModel, CustomListModel
-from mlte.store.base import StoreType
+from mlte.store.base import StoreType, StoreURI
+from mlte.store.custom_list.factory import create_custom_list_store
 from mlte.store.custom_list.store import CustomListStore
+from mlte.store.custom_list.underlying.fs import FileSystemCustomListStore
 from mlte.store.custom_list.underlying.http import HttpCustomListStore
+from mlte.store.custom_list.underlying.memory import InMemoryCustomListStore
+from mlte.store.custom_list.underlying.rdbs.store import RDBCustomListStore
 from mlte.user.model import ResourceType, UserWithPassword
-from test.backend.fixture import user_generator
-from test.backend.fixture.test_api import TestAPI
-from test.store.custom_list.custom_list_store_creators import (
-    create_fs_store,
-    create_http_store,
-    create_memory_store,
-    create_rdbs_store,
-)
+from test.store.defaults import IN_MEMORY_SQLITE_DB
+from test.store.fixture import create_api_and_http_uri
 
 CUSTOM_LIST_BASE_URI = f"{settings.API_PREFIX}/{ResourceType.CUSTOM_LIST.value}"
 """Base URI for custom lists."""
@@ -32,6 +32,31 @@ DEFAULT_ENTRY_DESCRIPTION = "test description"
 DEFAULT_PARENT = None
 
 
+def create_memory_store() -> InMemoryCustomListStore:
+    return typing.cast(
+        InMemoryCustomListStore,
+        create_custom_list_store(
+            StoreURI.create_uri_string(StoreType.LOCAL_MEMORY)
+        ),
+    )
+
+
+def create_fs_store(path: Path) -> FileSystemCustomListStore:
+    return typing.cast(
+        FileSystemCustomListStore,
+        create_custom_list_store(
+            StoreURI.create_uri_string(StoreType.LOCAL_FILESYSTEM, str(path))
+        ),
+    )
+
+
+def create_rdbs_store() -> RDBCustomListStore:
+    return RDBCustomListStore(
+        uri=StoreURI.from_string(IN_MEMORY_SQLITE_DB),
+        poolclass=StaticPool,
+    )
+
+
 def create_api_and_http_store(
     user: Optional[UserWithPassword] = None,
 ) -> HttpCustomListStore:
@@ -39,16 +64,8 @@ def create_api_and_http_store(
     Get a HttpStore configured with test client.
     :return: The configured store
     """
-    user = user_generator.build_admin_user()
-    test_api = TestAPI(user=user)
-    client = test_api.get_test_client()
-
-    return create_http_store(
-        username=client.username,
-        password=client.password,
-        uri=str(client.client.base_url),
-        client=client,
-    )
+    client, uri = create_api_and_http_uri(user)
+    return HttpCustomListStore(uri=uri, client=client)
 
 
 @pytest.fixture(scope="function")
