+++ test/store/user/fixture.py
@@ -9,9 +9,7 @@ from typing import Optional
 import pytest
 from sqlalchemy.pool import StaticPool
 
-from mlte._private import url as url_utils
 from mlte.store.base import StoreType, StoreURI
-from mlte.store.common.http_clients import OAuthHttpClient
 from mlte.store.user.factory import create_user_store
 from mlte.store.user.store import UserStore
 from mlte.store.user.underlying.fs import FileSystemUserStore
@@ -19,9 +17,8 @@ from mlte.store.user.underlying.http import HttpUserStore
 from mlte.store.user.underlying.memory import InMemoryUserStore
 from mlte.store.user.underlying.rdbs.store import RelationalDBUserStore
 from mlte.user.model import UserWithPassword
-from test.backend.fixture import user_generator
-from test.backend.fixture.test_api import TestAPI
-from test.store.defaults import IN_MEMORY_SQLITE_DB, get_http_defaults_if_needed
+from test.store.defaults import IN_MEMORY_SQLITE_DB
+from test.store.fixture import create_api_and_http_uri
 
 CACHED_DEFAULT_MEMORY_STORE: Optional[InMemoryUserStore] = None
 """Global, initial, in memory store, cached for faster testing."""
@@ -58,23 +55,6 @@ def create_rdbs_store() -> RelationalDBUserStore:
     )
 
 
-def create_http_store(
-    username: Optional[str] = None,
-    password: Optional[str] = None,
-    uri: Optional[str] = None,
-    client: Optional[OAuthHttpClient] = None,
-) -> HttpUserStore:
-    username, password, uri = get_http_defaults_if_needed(
-        username, password, uri
-    )
-    return HttpUserStore(
-        uri=StoreURI.from_string(
-            url_utils.set_url_username_password(uri, username, password)
-        ),
-        client=client,
-    )
-
-
 def create_api_and_http_store(
     user: Optional[UserWithPassword] = None,
 ) -> HttpUserStore:
@@ -82,16 +62,8 @@ def create_api_and_http_store(
     Get a HttpStore configured with test client.
     :return: The configured store
     """
-    user = user_generator.build_admin_user()
-    test_api = TestAPI(user=user)
-    client = test_api.get_test_client()
-
-    return create_http_store(
-        username=client.username,
-        password=client.password,
-        uri=str(client.client.base_url),
-        client=client,
-    )
+    client, uri = create_api_and_http_uri(user)
+    return HttpUserStore(uri=uri, client=client)
 
 
 @pytest.fixture(scope="function")
